code refactor. externalized stylesheet
