+++ App/actions/email.js
@@ -1,6 +1,8 @@
 var email = exports = module.exports
 
 // Actions dispatched from search page
+
+//adds email to email array
 exports.addEmail = function addEmail(email) {
   console.log(email)
   return {
@@ -9,6 +11,8 @@ exports.addEmail = function addEmail(email) {
   }
 }
 
+
+//truthy property on state while events API is hit
 exports.loadingPoll = function loadingPoll(bool) {
   console.log('load poll action', bool)
   return {
@@ -17,6 +21,7 @@ exports.loadingPoll = function loadingPoll(bool) {
   }
 }
 
+//removes email from email array
 exports.delEmail = function delEmail(emailAddress) {
   return {
     type: 'DEL_EMAIL',
@@ -24,6 +29,7 @@ exports.delEmail = function delEmail(emailAddress) {
   }
 }
 
+//array of contacts
 exports.addContacts = function addContacts(contacts){
   return {
     type: 'ADD_CONTACTS',
