+++ App/actions/login.js
@@ -1,7 +1,7 @@
 
 var login = exports = module.exports
 
-// Actions dispatched from search page
+// action sent out after response is received from facebook server regarding auth
 exports.user_set = function user_set(user_id, user_last_name, user_first_name, user_email) {
   return {
     type: 'USER_SET',
